"6/17 -10am"

+++ src/main/java/com/example/firstdraftspringfinalproject/models/Employee.java
@@ -15,7 +15,7 @@ public class Employee {
     private ArrayList<Timesheet> timesheets;
     private Boolean currentTimesheetCompletionStatus = true;
     private Boolean supervisorAccess;
-    private float payRate;
+    private Integer payRate;
     private GregorianCalendar firstDateOfWork;
 
     // Pay Rate?, eligible for benefits, remaining time off
